WL#12771 reuse connection [01/26] - classic::auth, classic::quit
Issue
=====

The protocol splicer currently operates on frame-level of the classic
and xprotocol. It knows where a frame starts and where it ends.

  frame:
    frame-length
    sequence-id
    frame-payload (opaque)

For the reuse of connections, the routing code needs to understand the
flow of the messages, and not just the frame around them.

  frame:
    frame-length
    sequence-id
    frame-payload (message)

To decode the frame-payload into a message, the state of the protocol
must be tracked: A field-count message of a resultset is followed by a
column-metadata messages.

To reuse the connections the routing code has to know what capabilities
the client supports which requires a change if the initial message flow.

Before,

1. the client connects to the router
2. router connects to a server
3. server sends server::greeting
4. router forwards server::greeting to client
5. client sends client::greeting to router

Now,

1. the client connects to the router
2. router sends server:greeting to client
3. client sends client::greeting to router
4. router checks if a connection exists matching the routers
   capabilities and connects to server
5. server sends server::greeting to router
6. router sends client::greeting to server

Change
======

- refactored blocked-endpoints from context.h into its own class
  and header-file
- refactored BasicSplicer and ProtocolSplicers into:
  - class TlsSwitchable
  - class ConnectionBase
  - class BasicConnection
  - class RoutingConnection
  - class ClassicProtocolState
  - class TlsSwitchableConnection
- moved ClassicProtocolState to its own header
- moved Connector from mysql_routing.cc to connection.h to allow
  lazy-connect
- moved intitial connection attributes into its own header
- disabled xprotocol support for now (added back later)
- moved "class Owner" into its own file
- added flow for authentication and COM_QUIT

RB: 27370