WL#14942 REFACTORS SOME EXISITING NDB REPLICATION TESTS
Some NDB replication tests which verify the conflict detection
and resolution algorithms supported use a secondary table to
demonstrate how the replication would behave if the table isn't
configured with conflict detection and resolution functions.

The schema of this secondary table is updated to closely match
the schema of the table which is configured with conflict
detection and resolution functions. Also, since the secondary
table doesn't generate exceptions while replicating, the
queries which look up the exception table for the secondary
table are removed.

Change-Id: Ib7db1695ba46eedd168b52d655f13248fa6d97b3