BUG#33283709 SHOW TABLES or information_schema.TABLES doesn't show tables if a user has an access privileges to individual performance schema tables
Problem:
MySQL Server lists tables by using SHOW TABLES command, or accessing information_schema.TABLES, and other variants, if the user has any privileges on a table. However, it's not true for the performance schema. SHOW TABLES or SELECT on information_schema doesn't list out the tables in this case, the output is empty. However, the user can perform SELECT query on the table directly. It works fine though if the user has access to every table in the performance_schema database. There are few tables which are known as world-only tables which are always visible to the user, even they're not showing up.

Root Cause:
The check methods inside of the pfs_engine_table.cc are responsible for checking the permissions for the user on the tables. The parameter named "want_access" is a bit field which tells the funtion what operations are requested by the user. Every check function has a bit field of forbidden operations within them, if want_access parameter has any bit turned on which is forbidden inside of the forbidden bit field, it'll return access denied to the user. But if the want_access parameter has any bit
turned on which is NOT in the forbidden ones, the function should never return access denied, instead, the allowed operations should be granted, and the function should return check for grants.

Solution:
This is happening because the check methods don't have enough parameters to properly check for the permissions, added one more parameter named "any_combination_will_do", and this parameter is true if it's enough to have any privilege for any combination of the table columns. Added this parameter in the check funtions, and modified the code accordingly. So, if this parameter is true, the function will check if there's any allowable bit in the want_access parameter, if it exists, grant those permissions and return. If any_combination_will_do is false, then the code works in the same way as it was working previously. The written tests basically checks if the world only tables are visible to the user initially, then, grants SELECT permission on few tables, and then perform SELECT again to see if both world only, as well as, those few tables are visible or not.

RB: 27554