Bug #33769505 Potentially incompatible system Boost gets picked even with -DWITH_BOOST=/path
On macOS, we may configure with "system" libraries from Homebrew.  If
boost is installed in Homebrew, we may include header files from
HOMEBREW_HOME/boost rather than from WITH_BOOST or our own
include/boost/patches. This may break the build and/or generate code
which breaks GIS functions.

The solution is to disable misc -DWITH_XXX=system if boost is found in
Homebrew.

Also ensure that we look for system headers in homebrew/include if any
homebrew libraries have been chosen in place of the bundled ones.

Change-Id: I4eecc192b701e1c746ea7e9493b16147ae4b463a