Bug#33772598 Replace EncryptionPasswordData with EncryptionKeyMaterial
Changing the name to reflect the new use also for passing keys not only
passwords.

Also moving it to FsOpenReq.hpp to which it is closely related.

Also removing some redundant null termination of password data, keeping
it only in managment server to not break compatibility with old data
nodes.

Since backups with mixed versions of data nodes are not allowed there is
no need to keep null termination for compatibility.

Change-Id: I44c3ec6f09d90fc7327ea51a7e3139aa6142dec7