Bug#33722819: recent regression: crash/corruption with gcols and multitable delete...
Post-push fix.

The original patch meant to call SetUpTablesForDelete() before the
child iterators of the DeleteRowsIterator was created, but it actually
calls it twice; both before and after the child iterators are created.
While this is harmless, it's doing unnecessary work. This patch
changes it so that it is only called once.

Change-Id: Ib68a1ff87fa32ae2701afdec47ac368388397029