WL#14672: Enable the hypergraph optimizer for UPDATE [4/8, use path]
Move the call to Query_result_update::optimize() later in the
optimization phase, just before the iterators are created. The main
point is that it is now called after the old optimizer has created the
access paths, so that a common code path can be used for the old
optimizer and the hypergraph optimizer. optimize() is also changed so
that it uses access paths where it can.

Before this patch, Query_result_update was the only class with a
non-trivial implementation of Query_result::optimize(). Now it is the
only implementation that is called. Because of this,
Query_result::optimize() is removed, and
Query_result_update::optimize() is made non-virtual.

Change-Id: Ib919f1bae300068a1b9bc649b44e374bcc1c5624