Bug #33784168 UnknownClusterSetTargetCluster/5 fails randomly
The test is taking a wrong assumption that having an empty target_cluster
for a Router in the metadata leads to an error and failure to route user connections.
In reality it is only a warning and the connections are handled. It sometimes
was passing because the synchronization point was missing after waiting for
specific warning message and the pass/fail depended on the timing (race
condition).

This patch adds the synchronization point to the existing test to wait for the
metadata to always be propagated in the Router and extracts the one testcase
that is not an error scenario (only warning).

Change-Id: Ibee6aa99098c2d065324844537adbc1956187f5c