Bug#33535379 Wrong value for RAND() expression when aggregating on empty table.
This error happens for queries with the following properties:
* The result set has a column X with a non-deterministic element.
* X does not refer any table.
* There is implicit grouping.
* The query plan is such that we aggregate from a temporary table.
* That temporary table is empty.

An example would be:

INSERT INTO tab SELECT MAX(tab.col), RAND(0) AS x FROM tab;

where 'tab' is empty. (By inserting into the same table, we ensure
that the query plan will aggregate from a temporary table.)
The temporary table will have a column for 'x', and the (single) result
row will pick 'x' from the first row of the temporary table. If the
temporary table is empty, there is no first row, and 'x' becomes 0.
So in the query above, '(NULL, 0)' rather than '(NULL, 0.155..)' will
be inserted in 'tab'.

The fix is to avoid storing such columns in the temporary table.
We keep the original Item_func, instead of replacing it with an
Item_field referring to the temporary table.

Change-Id: I3f133be06c4b480514c8381052ac3564cbd203d9