Bug#33542052 grant ndb_stored_user alters user with wrong hash
This patch changes ndb_stored_grants.cc to set the session variable
print_identified_with_as_hex=true, so that the password hash stored
in the ndb_sql_metadata table is formatted as a hex literal rather
than a string literal.

The bug report contains "error: 1827: The password hash doesn't have
the expected format", which suggests that the string was corrupted
at some point. These strings are known to have unprintable characters
that can cause corruption in some environments, which is why the
print_identified_with_as_hex option was introduced.

The included test case in ndb_rpl_stored_grants attempts to reproduce the
issue as described in the bug report, but I did not actually succeed at
reproducing the problem in any of my test environments.

Change-Id: I9c394da794a13f31a90a68c8d44a345b28b0d2d8