Bug#32929392 - TEMPTABLE STORAGE ENGINE LACKS OF THE ERROR HANDLER OF RESULT::RECORD_FILE_FULL
In case of a distinct aggregator function, the rows are collected into
a in-memory tree for duplicate elimination. If the aggregation is on a
LOB column a temporary table is used to hold the rows.

If the temp table storage engine exhausts memory and reports a file
full, the error is not handled and results in inconsistent behaviour.

Fix:
If the temporary table storage engine reports file full, move the
table to InnoDB and continue with the operation.

Change-Id: I7eb139d0deabfdc7494beb7fb1af5e866dbf9948