Bug#33778923 Collection of 'EventBytesRecvdCount' client statistics has too much overhead
The method EventBufData::get_size() implements the size calculation of a
collection of EventBufData's by traversing the m_next and m_next_blob
pointers available through the head-EventBufData. Profiling has
shown a rather large overhead related to this operation. It is likely
contributing to trashing the L1,L2 and L3 caches as well, as huge amount
of data might get visited. It will be better to just collect
this 'size' and 'count' information when we construct the EventBufData
in the first place.

This patch introduce the class EventBufDataHead, a subclass of EventBufData
for aggregation of EventBufData sizes.

The class EventBufDataHead is introduced as a subclass of EventBufData.
It is used to represent the 'main_data' EventBufData, which is the 'head'
of a set of EventBufData which will be returned by ::pollEvents().

The EventBufDataHead extends the EventBufData class by also maintain
an aggregate of the total count and size of the EventBufData linked
to the 'Head'. We use this aggregated size, instead of get_size(),
when updating the Ndb client statistics.

Change-Id: I3fecafa927cc7e06136f2efc986b580d0c0ec1f1
(cherry picked from commit db089ce58b8fefb94cf653910f206f82718f633a)