Bug#33791802: Query with multiple JOINs stopped working
A query having a natural join involving a merged view
results in errors while pushing down a condition to
the underlying derived table of the merged view.
When a natural join is setup, it changes the context
of the query block to point to the join list. The same
context is used to resolve the field in the where clause.
During simplification of the joins, the join list starts
pointing to the underlying tables/derived tables of the
merged view. Since condition pushdown happens after
simplification of joins, the context of the where condition
for resolving does not work anymore because we have a field
which needs to be resolved in the context of the merged
view not the underlying tables/derived tables.

So the solution is to use the underlying expression for condition
pushdown in case of a view reference instead of the
reference itself. Before WL#13730, a query block's context
was used to resolve any expression during condition pushdown.
So we needed to use the view reference so that it resolved
successfully in the context of the derived table's query block.
However with WL#13730, we collect the contexts of the fields that are
part of the expression and set it for the cloned expression.
This means, we are not dependent on the context of the query block
anymore. As a result, we can now use the underlying expression in case
of a view reference since we will be using the underlying expression's
context itself to resolve the expression later. This also means, we
do not need special handling for view references while collecting
contexts. Additionally, we no more need the special QT flag for
printing which was used mainly for view reference, extended later
for printing a sp variable (We can use "reparse_derived_condition"
for the later).

Change-Id: I962319b0161866b9469197fa6590a2babfed5785