Bug #33783709 mem_heap_alloc could provide better hints to VALGRIND
The `UNIV_DEBUG` variant of `mem_heap_alloc` adds `MEM_NO_MANS_LAND` (=16) bytes on each side of the allocated fragment of memory and uses these margins to detect errornous writes outside of the allocated fragment.
Additionally, the `UNIV_DEBUG_VALGRIND` will mark both of these margins with `VALGRIND_MAKE_MEM_NOACCESS` macro, so that when run with Valgrind, we will detect not only writes, but also reads which try to access the margins.

However, `mem_heap_alloc` needs to ensure that the returned address is aligned to `UNIV_MEM_ALIGNMENT` (=8), and as it serves allocation requests by returning consecutive fragments of the same block, it achieves this by simply rounding the requested size and margins (`N + 2 * MEM_NO_MANS_LAND`) to the nearest multiple `UNIV_MEM_ALIGNMENT` - this way, by induction, each returned pointer is aligned.

The problem is that we mark only the last `MEM_NO_MANS_LAND` bytes of this rounded-up region with `VALGRIND_MAKE_MEM_NOACCESS`, which means that the padding which was added to round-up to multiple of `UNIV_MEM_ALIGNMENT` is still accessible and sits right after the buffer. That means, off-by-one errors will not be detected in case `N` is not a multiple of 8.

This patch fixes this by marking all the bytes after the requested `N`-bytes region with `VALGRIND_MAKE_MEM_NOACCESS`.

Reviewed by: Paweł Olchawa <pawel.olchawa@oracle.com>
RB:27614

Change-Id: I384ca2129e1dd84955fd16410979df0c2150854c