Bug#33806018: Remove usage of MD5 in explain_access_path.cc
In debug builds only, the hypergraph optimizer has functionality for
forcing certain subplans to be preferred, regardless of their cost.
The identifiers for these subplans are created from MD5 checksums.

Since MD5 is being phased out, we change it to create identifiers from
SHA-256 checksums instead, so that it will continue to work also when
MD5 is not available.

Change-Id: I243959734debeef78d295523b046c2314fcfc931