BUG#33302400 SYS.SCHEMA_UNUSED_INDEXES IS NOT ACCURATE AND NOT UPDATING
Description: sys.schema_unused_indexes does not update when the
indexes are being used to fetch from the table for the index values
that are not found in the table.

Analysis: schema_unused_indexes lists the indexes that are not used
by fetching it from the table_io_waits_summary_by_index_usage table
which has count_star as 0. Count_star is incremented only when the
index is used to query a row that is present in the table, while
the row is not found in the table count_star is not incremented.
inspite of usage of index scan which returns empty row it should
be removed from schema_unused_indexes.

Fix: Fix is added to update the sys.schema_unused_indexes properly
even if the row is not found in the table. This is fixed to
increment the read/fetch count of an index regardless of the
result except the ha_err_end_of_file.

RB: 27615