Bug#33813951: Binary resultset for `SELECT ?` with date/time gives wrong type
This is a regression from bug#32915973. Before this bugfix, with
a dynamic parameter that was derived as VARCHAR, and a parameter value
with a temporal actual type was supplied, the query was reprepared so
that the parameter was given the actual data type. This logic was lost
in that bugfix.

The solution here is to reinstate that logic. We have also added C API
test cases to validate that correct data types are returned.

Change-Id: I2448e3f2f36925fe70d882ae5681a6234f0d5a98