BUG#33824392: Improve show_rpl_debug_info.inc
Small improvements to show_rpl_debug_info.inc

* Background

show_rpl_debug_info.inc is sometimes used by tests to output debug
info after a failure. It was written long ago and would benefit from
some maintenance now.

* Changes

- When the file was written, mtr's loop feature only supported loops
that iterate in decreasing order. Now it can iterate in increasing
order, meaning we can make the output appear in a more natural order.

- When the file was written, there was no easy way to show the error
log. Now there is performance_schema.error_log. So we print it.

- mtr echoes the last 30 lines of the test log to its stdout.
Sometimes the full log is deleted and the 30 lines is all information
we get.  So it would help to put some of the more critical information
there, in particular, any errors in the replication threads and the
last errors from the error logs.

- Use new terminology.

- Improve formatting.

Change-Id: I64629e5ab2eb89ee06b68518aabf22b9f86810cd